Add basic ECDH functionality.
I encountered a bunch of problems getting this to work, and I'm still
not sure what resolved them exactly.

I find Cabal's docs to be confusing as to where to place the
-DENABLE_MODULE_ECDH option (it seems to state that it should be under
'cc-options'), but GHC suggests placing it under 'cpp-options', where I
had originally put it.

I originally encountered dyld errors around lazy symbol loading when
trying to call secp256k1's ECDH function in a REPL; the solution may
have been to simply build the package with cabal prior to using the
REPL, as swapping the flags and includes around didn't seem to solve
anything.